fix: moved the last of the parsers away from the stdlib
